# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Citas Médicas",
        "description": "Documentación de la API",
        "contact": {
            "email": "contacto@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/listarPacientes": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Listar pacientes",
                "operationId": "651f00a5a749582544f14bb006822fc6",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Paciente"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Prohibido (rol requerido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/AgregarPaciente": {
            "post": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Crear paciente",
                "operationId": "58215dc33363d1788812e52f784eb520",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "documento",
                                    "nombre",
                                    "apellido",
                                    "email"
                                ],
                                "properties": {
                                    "documento": {
                                        "type": "string",
                                        "example": "1012345678"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "María"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Gómez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3001234567",
                                        "nullable": true
                                    },
                                    "genero": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F",
                                            "O"
                                        ],
                                        "example": "F",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paciente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "403": {
                        "description": "Prohibido (rol requerido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/paciente/{id}": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Obtener paciente por ID",
                "operationId": "b5b291bbb69a7e21bd2eadac0731736d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paciente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "403": {
                        "description": "Prohibido (rol requerido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ActualizarPaciente/{id}": {
            "put": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Actualizar paciente",
                "operationId": "80a2ac8fc216e78eb0925ef9bcba786b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "documento": {
                                        "type": "string",
                                        "example": "1012345678"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "María"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Gómez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3001234567",
                                        "nullable": true
                                    },
                                    "genero": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F",
                                            "O"
                                        ],
                                        "example": "F",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paciente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "403": {
                        "description": "Prohibido (rol requerido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/EliminarPaciente/{id}": {
            "delete": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Eliminar paciente",
                "operationId": "836b9fcd2361022c9febcad091a3c7b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Paciente eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "403": {
                        "description": "Prohibido (rol requerido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pacientesMasculinos": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Listar pacientes masculinos",
                "operationId": "eaad92dcccb3a19fd2e9570a75480473",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Paciente"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Prohibido (rol requerido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Paciente": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "documento": {
                        "type": "string",
                        "example": "1012345678"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "María"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Gómez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "maria@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "3001234567",
                        "nullable": true
                    },
                    "genero": {
                        "type": "string",
                        "enum": [
                            "M",
                            "F",
                            "O"
                        ],
                        "example": "F",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}